+++ app/src/main/res/layout/item_grade_statistics_pie.xml
@@ -1,0 +1,8 @@
+<com.github.mikephil.charting.charts.PieChart xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/gradeStatisticsChart"
+    android:layout_width="match_parent"
+    android:layout_height="400dp"
+    android:layout_margin="10dp"
+    android:background="?android:windowBackground"
+    tools:context=".ui.modules.grade.statistics.GradeStatisticsAdapter" />
